+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Authentication - Improve the log message when 2FA secret key is not set.
 - Technical - Use stubs instead of object dependencies for test purposes.
+- Technical - Change ESLint ruleset for Airbnb.
 
 ### Fixed
 - Technical - Make the tests run on the non-transpiled sources.